churn: blackened xcorr

+++ data/timescale-initdb.sql
@@ -8,7 +8,7 @@ CREATE TABLE recordings (
   sensor_id       INT               NOT NULL,
   sensor_type_id  INT               NOT NULL,
   sample_count    INT               NOT NULL,
-  sample_rate     DOUBLE PRECISION  NOT NULL,
+  sample_rate     REAL              NOT NULL,
   x               REAL ARRAY        NULL,
   y               REAL ARRAY        NULL,
   z               REAL ARRAY        NULL
